+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.9 (2020-04-26)
+### Fixed bugs:
+- Fixed issue caused in v1.1.8 with running the `monitor` task in Windows Task Scheduler where the task would quit with an error code of 0xE0434352. [#13](https://github.com/kmahon37/plex-dvr-waker/issues/13)
+
+### Other:
+- Added some logging improvements that should help with tracking down future issues.
+
 ## 1.1.8 (2020-04-19)
 ### Fixed bugs:
 - Fixed issue when Windows Task Scheduler cannot automatically find `dotnet.exe` on some systems like Windows 7 (and maybe others).  It will now check for `dotnet.exe` at the default install location of `%ProgramFiles%\dotnet\dotnet.exe`, and if not found then it will also check all locations in the `PATH` environment variable.  If `dotnet.exe` cannot be found, it will error out when running one of the `add-task` commands. [#13](https://github.com/kmahon37/plex-dvr-waker/issues/13)